+++ app/Http/Controllers/Api/Partner/V1/PartnerController.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace App\Http\Controllers\Api\Admin\V1;
+namespace App\Http\Controllers\Api\Partner\V1;
 
 use App\Dtos\PartnerDto;
 use App\Http\Controllers\Controller;
-use App\JsonApi\Admin\V1\Partners\PartnerQuery;
-use App\JsonApi\Admin\V1\Partners\PartnerRequest;
-use App\JsonApi\Admin\V1\Partners\PartnerSchema;
+use App\JsonApi\Partner\V1\Partners\PartnerQuery;
+use App\JsonApi\Partner\V1\Partners\PartnerRequest;
+use App\JsonApi\Partner\V1\Partners\PartnerSchema;
 use App\Models\Partner;
+use App\JsonApi\Proxies\PartnerPartner as PartnerPartner; // proxy model https://laraveljsonapi.io/docs/1.0/digging-deeper/proxies.html
 use App\Services\PartnerService;
 use LaravelJsonApi\Core\Document\Error;
 use LaravelJsonApi\Core\Responses\DataResponse;
@@ -24,9 +25,9 @@ class PartnerController extends Controller
 //    use Actions\Destroy;
     use Actions\FetchRelated;
     use Actions\FetchRelationship;
-    use Actions\UpdateRelationship;
-    use Actions\AttachRelationship;
-    use Actions\DetachRelationship;
+//    use Actions\UpdateRelationship;
+//    use Actions\AttachRelationship;
+//    use Actions\DetachRelationship;
 
     /**
      * @var PartnerService
@@ -46,6 +47,7 @@ class PartnerController extends Controller
      * @param PartnerQuery $query
      * @return false|DataResponse
      */
+    /*
     public function store(PartnerSchema $schema, PartnerRequest $request, PartnerQuery $query)
     {
         $attributes = $request->data['attributes'];
@@ -67,26 +69,33 @@ class PartnerController extends Controller
         }
 
         $partner = Partner::find($partner->getKey());
+        $partner = new PartnerPartner($partner);
         return new DataResponse($partner);
     }
+    */
 
     /**
      * Update partner
      * @param PartnerSchema $schema
      * @param PartnerRequest $request
      * @param PartnerQuery $query
-     * @param Partner $partner
+     * @param PartnerPartner $partner
      * @return false|DataResponse
      */
 
-    public function update(PartnerSchema $schema, PartnerRequest $request, PartnerQuery $query, Partner $partner)
+    //public function update(PartnerSchema $schema, PartnerRequest $request, PartnerQuery $query, PartnerPartner $partner)
+    public function update(PartnerSchema $schema, PartnerRequest $request, PartnerQuery $query, PartnerPartner $partner)
+    //public function update()
     {
+        //dd($partner);
+        print 'test';
+        exit;
         $attributes = $request->data['attributes'];
 
         $dto = new PartnerDto(
             $attributes['name'] ?? $partner->name,
             $attributes['info'] ?? $partner->info,
-            $attributes['isBlock'] ?? $partner->is_block,
+            $partner->is_block, // пользователь-партнер не может менять состояние блокировки
             $partner->phone, //$attributes['phone'],  Запрещено менять номер телефона
         );
 
@@ -100,6 +109,7 @@ class PartnerController extends Controller
         }
 
         $partner = Partner::find($partner->getKey());
+        $partner = new PartnerPartner($partner);
         return new DataResponse($partner);
     }
 
@@ -107,10 +117,10 @@ class PartnerController extends Controller
      * Удаление существующего ресурса. Замена на блокировку партнера.
      *
      * @param PartnerRequest $request
-     * @param Partner $partner
+     * @param PartnerPartner $partner
      * @return \Illuminate\Contracts\Support\Responsable|\Illuminate\Http\Response
      */
-    public function destroy(PartnerRequest $request, Partner $partner)
+    public function destroy(PartnerRequest $request, PartnerPartner $partner)
     {
         //var_dump($user->roles());
         //exit;
@@ -131,6 +141,7 @@ class PartnerController extends Controller
         }
 
         $partner = Partner::find($partner->getKey());
+        $partner = new PartnerPartner($partner);
         return new DataResponse($partner);
     }
 }