+++ app/Models/PartnerUser.php
@@ -7,6 +7,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Database\Eloquent\Relations\Pivot;
 
 /**
  * Class PartnerUser
@@ -27,7 +28,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  *
  * @package App/Models
  */
-class PartnerUser extends Model
+class PartnerUser extends Pivot
 {
     use HasFactory, SoftDeletes;