+++ routes/api.php
@@ -12,13 +12,17 @@ use App\Http\Controllers\Api\Admin\V1\PartnerController;
 use App\Http\Controllers\Api\Admin\V1\CompanyController;
 use App\Http\Controllers\Api\Admin\V1\CompanyUserController;
 use \App\Http\Controllers\Api\Admin\V1\PartnerUserController;
+use \App\Http\Controllers\Api\Admin\V1\AccountController;
 
 //Controller of customer
 use App\Http\Controllers\Api\Customer\V1\CompanyUserContorller as CustomerCompanyUserController;
+use \App\Http\Controllers\Api\Customer\V1\AccountController as CustomerAccountController;
 
 //Partner Controllers
 use \App\Http\Controllers\Api\Partner\V1\EmployeeController as PartnerEmployeeController;
 use \App\Http\Controllers\Api\Partner\V1\CompanyController as PartnerCompanyController;
+use \App\Http\Controllers\Api\Partner\V1\PartnerController as PartnerPartnerController;
+use \App\Http\Controllers\Api\Partner\V1\AccountController as PartnerAccountController;
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -30,19 +34,19 @@ use \App\Http\Controllers\Api\Partner\V1\CompanyController as PartnerCompanyCont
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
+/*Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
-});
+});*/
 
 // Route::post('auth/phone', [AuthController::class, 'login']);
 // Route::post('auth/phone/{phone}', [AuthController::class, 'checkCode'])->whereNumber('phone');
 
-Route::group(['middleware' => 'auth:sanctum'], function () {
-    Route::post('auth/logout', [AuthController::class, 'logout']); // logout
+//Route::group(['middleware' => 'auth:sanctum'], function () {
+    //Route::post('auth/logout', [AuthController::class, 'logout']); // logout
 
     // User routes
-    Route::resource('users', UserController::class); // не json-api спецификация!
-});
+    //Route::resource('users', UserController::class); // не json-api спецификация!
+//});
 
 Route::prefix('admin/v1')
     ->group(function(){
@@ -50,9 +54,9 @@ Route::prefix('admin/v1')
     Route::post('/auth', [AuthController::class, 'loginAdmin']);
     //Route::post('auth/phone/{phone}', [AuthController::class, 'checkCodeAdmin'])->whereNumber('phone');
     Route::post('/auth/{phone}', [AuthController::class, 'checkCodeAdmin'])->whereNumber('phone');
-    Route::group(['namespace'=>'Api/Admin/V1'], function (){
+    /*Route::group(['namespace'=>'Api/Admin/V1'], function (){
         Route::post('/auth2/{phone}', [UserForAdminController::class, 'checkCode'])->whereNumber('phone');
-    });
+    });*/
 
     });
 
@@ -74,7 +78,15 @@ JsonApiRoute::server('Admin\V1')
     ->prefix('admin/v1')
     ->middleware('auth:sanctum')
     ->resources(function ($server) {
-        Route::post('auth/logout', [AuthController::class, 'logout']); // logout
+        //Route::post('auth/logout', [AuthController::class, 'logout']); // logout
+
+        $server->resource('account', AccountController::class);
+        Route::delete('account', [AccountController::class, 'logout']); // logout
+        /**
+         * POST     /account - изменить аккаунт
+         * GET      /account - посмотреть данные по аккаунту
+         * DELETE  /account  - выход из системы
+         */
 
         //roles routes
         $server->resource('roles', RoleController::class)
@@ -98,8 +110,8 @@ JsonApiRoute::server('Admin\V1')
         //companies
         $server->resource('companies', CompanyController::class)
             ->relationships(function($relationships) {
-               $relationships->hasMany('companyUsers'); // только просмотр
-               $relationships->hasOne('owner')->only('related', 'show');;
+               $relationships->hasMany('companyUsers'); // только просмотр // ?
+               $relationships->hasOne('owner')->only('related', 'show'); // ?
             });
 
         //company users
@@ -108,8 +120,8 @@ JsonApiRoute::server('Admin\V1')
         //partners routes
         $server->resource('partners', PartnerController::class)
             ->relationships(function($relationships) {
-                $relationships->hasMany('partnerUsers'); // только просмотр
-                $relationships->hasOne('owner')->only('related', 'show');;
+                $relationships->hasMany('partnerUsers'); // только просмотр  // ?
+                $relationships->hasOne('owner')->only('related', 'show'); // ?
             });
 
         //partner users routes
@@ -122,9 +134,13 @@ JsonApiRoute::server('Customer\V1')
     ->prefix('customer/v1')
     ->middleware('auth:sanctum')
     ->resources(function ($server) {
-        Route::post('auth/logout', [AuthController::class, 'logout']); // logout
+        //Route::post('auth/logout', [AuthController::class, 'logout']); // logout
+        $server->resource('account', CustomerAccountController::class);
+        Route::delete('account', [CustomerAccountController::class, 'logout']); // logout
         // company user routes
-        $server->resource('company-users', CustomerCompanyUserController::class);
+        $server->resource('company-users', CustomerCompanyUserController::class); // => employees
+        // partners
+        // companies
     });
 
 
@@ -134,10 +150,15 @@ JsonApiRoute::server('Partner\V1')
     ->prefix('partner/v1')
     ->middleware('auth:sanctum')
     ->resources(function ($server) {
-        Route::post('auth/logout', [AuthController::class, 'logout']); // logout
+        //Route::post('auth/logout', [AuthController::class, 'logout']); // logout
         // company user routes
         //$server->resource('companies', JsonApiController::class);
         $server->resource('companies', PartnerCompanyController::class);
+        $server->resource('account', PartnerAccountController::class);
+        Route::delete('account', [PartnerAccountController::class, 'logout']); // logout
+        $server->resource('partners', PartnerPartnerController::class);
+        // customers
+        // my-companies , т.е. это partners
         $server->resource('employees', PartnerEmployeeController::class);
         // $server->resource('customers', EmployeeController::class); // Покупатели, реализовать в следующих релизах
     });